Factor out large component to separate file
