+++ src/components/income/OtherInvestments.tsx
@@ -28,21 +28,19 @@ interface Show<A> {
   (a: A): string
 }
 
-const toShow = <A,>(f: (a: A) => string): Show<A> => f
-
 const show =
   <A,>(shows: Show<A>) =>
   (a: A): string =>
     shows(a)
 
-const showAssetType: Show<AssetType> = toShow((p) => {
+const showAssetType: Show<AssetType> = (p) => {
   switch (p) {
     case 'Security':
       return 'Security (Stock, bond, option, mutual fund, etc.)'
     case 'Real Estate':
       return 'Real Estate'
   }
-})
+}
 
 interface AssetUserInput {
   name: string
